fix: Improve SPDX license scan reports for missing details
Currently, the `The license information in recipe` section is reliably
output. However, the sections for `The scan result got from spdx file (the same files as recipe)`
and `According to spdx file, there are some other files that maybe include license information. Such as:`
occasionally fail to output as intended.

This commit addresses these output inconsistencies to ensure all expected
license information is correctly presented in the reports.

Signed-off-by: Yoshihisa Morizumi <yoshi.morizumi@fujitsu.com>